Copied outdoor scene into indoor scene.
Added a new TeleportDoor script so player teleports instead of changing scenes. Copied the outdoor scene and placed it in the indoor scene. Doors now teleport the player instead of changing the scene.

Fixed an error where items in the overworld cannot be interacted with again after the initial interaction due to the interaction zone being destroyed.