Added audio assets to the game.
Added 3 sound effects (mouse click, door open and door close) and 2 music assets (home and outdoor) to the game.

Partially hard coded but made it so music changes depending on the area you are in. Sound effects now play when clicking through dialogue and leaving an area through a door.